created single plant species component that renders in different pages using id as prop
